rename progress_dot.jsx to ProgressDot
