+++ apps/src/templates/progress/ProgressBubble.jsx
@@ -3,9 +3,9 @@ import Radium from 'radium';
 import color from "@cdo/apps/util/color";
 
 // TODO - not sure I like these bubble color styles being keyed by these strings.
-// In a future PR, share with progress_dot, likely by using an Enum of different
+// In a future PR, share with ProgressDot, likely by using an Enum of different
 // allowable values.
-import { BUBBLE_COLORS } from '@cdo/apps/code-studio/components/progress/progress_dot';
+import { BUBBLE_COLORS } from '@cdo/apps/code-studio/components/progress/ProgressDot';
 
 export const DOT_SIZE = 30;
